functioning login/register/verification
Written in PHP, HTML, CSS, Javascript, MySQL. Website is able to login, register, and verify by sending an email to the user's email address with a generated verification code. Biggest Issues are that we'll need to get an online server up and running, code uses MySQL which we'll need to convert the queries into firebase, and input needs server and client-side sanitization.